requirements.txt
Flask==2.3.3
fpdf==1.7.2
google-generativeai==0.3.2
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1

